# apps/backend/docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/customers/me/addresses": {
      "post": {
        "operationId": "AddressController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAddressDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressDto"
                }
              }
            }
          },
          "404": {
            "description": "No customer profile yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add an address — the customer's first address always becomes default",
        "tags": [
          "Customer"
        ]
      },
      "get": {
        "operationId": "AddressController_list_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AddressDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List addresses (default first, then oldest first)",
        "tags": [
          "Customer"
        ]
      }
    },
    "/api/v1/customers/me/addresses/{id}": {
      "get": {
        "operationId": "AddressController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get one address",
        "tags": [
          "Customer"
        ]
      },
      "patch": {
        "operationId": "AddressController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAddressDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update an address",
        "tags": [
          "Customer"
        ]
      },
      "delete": {
        "operationId": "AddressController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete an address",
        "tags": [
          "Customer"
        ]
      }
    },
...
    "/api/v1/geography/states": {
      "post": {
        "operationId": "StateController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StateDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a state",
        "tags": [
          "Geography"
        ]
      },
      "get": {
        "operationId": "StateController_findAll_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StateDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all states (active and inactive)",
        "tags": [
          "Geography"
        ]
      }
    },
    "/api/v1/geography/states/{id}": {
      "get": {
        "operationId": "StateController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StateDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a state by id",
        "tags": [
          "Geography"
        ]
      },
      "patch": {
        "operationId": "StateController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StateDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a state — isActive is the only deactivation path",
        "tags": [
          "Geography"
        ]
      }
    },
...
    "/api/v1/availability/working-hours/me": {
      "post": {
        "operationId": "WorkingHoursController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkingHoursDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkingHoursDto"
                }
              }
            }
          },
          "400": {
            "description": "startTime must be before endTime",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add a recurring weekly working-hours window (a day may have several, e.g. a split shift)",
        "tags": [
          "Availability"
        ]
      },
      "get": {
        "operationId": "WorkingHoursController_listOwn_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkingHoursDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the current provider's working-hours windows",
        "tags": [
          "Availability"
        ]
      }
    },
    "/api/v1/availability/working-hours/me/{id}": {
      "patch": {
        "operationId": "WorkingHoursController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkingHoursDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkingHoursDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a working-hours window — isActive can pause it without deleting it",
        "tags": [
          "Availability"
        ]
      },
      "delete": {
        "operationId": "WorkingHoursController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a working-hours window",
        "tags": [
          "Availability"
        ]
      }
    },
...
  },
  "info": {
    "title": "SevaMitra API",
    "description": "SevaMitra backend API — see docs/ARCHITECTURE.md",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorDetailDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "BAD_REQUEST"
          },
          "message": {
            "type": "string",
            "example": "Incorrect OTP"
          },
          "details": {
            "type": "object"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetailDto"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "CreateAddressDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "enum": [
              "HOME",
              "WORK",
              "OTHER"
            ],
            "default": "HOME"
          },
          "line1": {
            "type": "string",
            "example": "221B Temple Street"
          },
          "line2": {
            "type": "string",
            "example": "2nd Cross"
          },
          "landmark": {
            "type": "string",
            "example": "Opposite bus stand"
          },
          "town": {
            "type": "string",
            "example": "Mysuru"
          },
          "district": {
            "type": "string",
            "example": "Mysuru"
          },
          "state": {
            "type": "string",
            "example": "Karnataka"
          },
          "pincode": {
            "type": "string",
            "example": "570001",
            "description": "6-digit Indian PIN code"
          },
          "latitude": {
            "type": "number",
            "example": 12.2958
          },
          "longitude": {
            "type": "number",
            "example": 76.6394
          },
          "isDefault": {
            "type": "boolean",
            "default": false,
            "description": "The first address a customer adds always becomes default regardless of this flag"
          }
        },
        "required": [
          "line1",
          "town",
          "pincode"
        ]
      },
      "AddressDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "customerId": {
            "type": "string",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "label": {
            "type": "string",
            "enum": [
              "HOME",
              "WORK",
              "OTHER"
            ],
            "example": "HOME"
          },
          "line1": {
            "type": "string",
            "example": "221B Temple Street"
          },
          "line2": {
            "type": "object",
            "example": "2nd Cross",
            "nullable": true
          },
          "landmark": {
            "type": "object",
            "example": "Opposite bus stand",
            "nullable": true
          },
          "town": {
            "type": "string",
            "example": "Mysuru"
          },
          "district": {
            "type": "object",
            "example": "Mysuru",
            "nullable": true
          },
          "state": {
            "type": "object",
            "example": "Karnataka",
            "nullable": true
          },
          "pincode": {
            "type": "string",
            "example": "570001"
          },
          "latitude": {
            "type": "object",
            "example": 12.2958,
            "nullable": true
          },
          "longitude": {
            "type": "object",
            "example": 76.6394,
            "nullable": true
          },
          "isDefault": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "customerId",
          "label",
          "line1",
          "town",
          "pincode",
          "isDefault",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateAddressDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "enum": [
              "HOME",
              "WORK",
              "OTHER"
            ],
            "default": "HOME"
          },
          "line1": {
            "type": "string",
            "example": "221B Temple Street"
          },
          "line2": {
            "type": "string",
            "example": "2nd Cross"
          },
          "landmark": {
            "type": "string",
            "example": "Opposite bus stand"
          },
          "town": {
            "type": "string",
            "example": "Mysuru"
          },
          "district": {
            "type": "string",
            "example": "Mysuru"
          },
          "state": {
            "type": "string",
            "example": "Karnataka"
          },
          "pincode": {
            "type": "string",
            "example": "570001",
            "description": "6-digit Indian PIN code"
          },
          "latitude": {
            "type": "number",
            "example": 12.2958
          },
          "longitude": {
            "type": "number",
            "example": 76.6394
          },
          "isDefault": {
            "type": "boolean",
            "default": false,
            "description": "The first address a customer adds always becomes default regardless of this flag"
          }
        }
      },
...
      "CreateStateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Karnataka"
          },
          "code": {
            "type": "string",
            "example": "KA"
          }
        },
        "required": [
          "name"
        ]
      },
      "StateDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "name": {
            "type": "string",
            "example": "Karnataka"
          },
          "code": {
            "type": "object",
            "example": "KA",
            "nullable": true
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateStateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Karnataka"
          },
          "code": {
            "type": "string",
            "example": "KA"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
...
      "CreateWorkingHoursDto": {
        "type": "object",
        "properties": {
          "dayOfWeek": {
            "type": "string",
            "enum": [
              "MONDAY",
              "TUESDAY",
              "WEDNESDAY",
              "THURSDAY",
              "FRIDAY",
              "SATURDAY",
              "SUNDAY"
            ],
            "example": "MONDAY"
          },
          "startTime": {
            "type": "string",
            "example": "09:00"
          },
          "endTime": {
            "type": "string",
            "example": "18:00"
          }
        },
        "required": [
          "dayOfWeek",
          "startTime",
          "endTime"
        ]
      },
      "WorkingHoursDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "providerId": {
            "type": "string",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "dayOfWeek": {
            "type": "string",
            "enum": [
              "MONDAY",
              "TUESDAY",
              "WEDNESDAY",
              "THURSDAY",
              "FRIDAY",
              "SATURDAY",
              "SUNDAY"
            ],
            "example": "MONDAY"
          },
          "startTime": {
            "type": "string",
            "example": "09:00"
          },
          "endTime": {
            "type": "string",
            "example": "18:00"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "providerId",
          "dayOfWeek",
          "startTime",
          "endTime",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateWorkingHoursDto": {
        "type": "object",
        "properties": {
          "dayOfWeek": {
            "type": "string",
            "enum": [
              "MONDAY",
              "TUESDAY",
              "WEDNESDAY",
              "THURSDAY",
              "FRIDAY",
              "SATURDAY",
              "SUNDAY"
            ],
            "example": "MONDAY"
          },
          "startTime": {
            "type": "string",
            "example": "09:00"
          },
          "endTime": {
            "type": "string",
            "example": "18:00"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}