# apps/backend/src/geography/dto/update-state.dto.ts
import { ApiProperty, PartialType } from '@nestjs/swagger';
import { IsBoolean, IsOptional } from 'class-validator';
import { CreateStateDto } from './create-state.dto';

/** isActive is the only deactivation path — there is no delete endpoint. */
export class UpdateStateDto extends PartialType(CreateStateDto) {
  @ApiProperty({ required: false })
  @IsOptional()
  @IsBoolean()
  isActive?: boolean;
}